updated core scripts and library structure
this restructure was centered around the new json i/o available
with the new Frew build not being compatible with the nested
class structure of the old frewpy. The main reason of this is
that with the json file imported as a python object, objects
manipulated in nested classes could not be passed up to the
parent class and then distributed amongst all of the nested
classes. the only solution around this is to open and close
the original json file every time a class method is run.
As this isnt reasonable its been restructured in a modular
way where the FrewModel just ties in all of the functions
to produce a useful means of dealing with the Frew model using
OOP. This also opens up the core of frewpy to be extended and
built upon by importing the core modules held within the models
root and used to solve whatever problem the user might have.
D
Development is still work in progress converting the old frewpy
to the new core structure and so commented out functions/methods
will be present